+++ source/win-about/App.vue
@@ -8,7 +8,6 @@
     v-bind:tabbar-label="windowTitle"
     v-bind:disable-vibrancy="true"
     v-on:tab="currentTab = $event"
-    v-on:statusbar-click="handleClick($event)"
   >
     <!--
       To comply with ARIA, we have to wrap the form in a tab container because
@@ -37,7 +36,7 @@
   </WindowChrome>
 </template>
 
-<script>
+<script lang="ts">
 /**
  * @ignore
  * BEGIN HEADER
@@ -54,6 +53,7 @@
 
 import WindowChrome from '../common/vue/window/Chrome.vue'
 import { trans } from '../common/i18n-renderer'
+import { defineComponent } from 'vue'
 
 // Import the tabs
 import GeneralTab from './General-Tab.vue'
@@ -64,7 +64,7 @@ import LicenseTab from './License-Tab.vue'
 import FontLicenseTab from './Font-License-Tab.vue'
 import DebugTab from './Debug-Tab.vue'
 
-export default {
+export default defineComponent({
   components: {
     WindowChrome,
     GeneralTab,
@@ -125,15 +125,15 @@ export default {
     }
   },
   computed: {
-    windowTitle: function () {
+    windowTitle: function (): string {
       if (process.platform === 'darwin') {
         return this.tabs[this.currentTab].label
       } else {
-        return trans('dialog.about.title') + ' ' + global.config.get('version')
+        return trans('dialog.about.title') + ' ' + (global as any).config.get('version')
       }
     }
   }
-}
+})
 </script>
 
 <style lang="less">