+++ source/win-about/Projects-Tab.vue
@@ -111,7 +111,7 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
 /**
  * @ignore
  * BEGIN HEADER
@@ -128,15 +128,16 @@
  */
 
 import { trans } from '../common/i18n-renderer'
+import { defineComponent } from 'vue'
 
-export default {
+export default defineComponent({
   name: 'ProjectsTab',
   data: function () {
     return {
       projectsIntro: trans('dialog.about.other_projects')
     }
   }
-}
+})
 </script>
 
 <style lang="less">