+++ source/win-stats/CalendarView.vue
@@ -65,7 +65,7 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
 /**
  * @ignore
  * BEGIN HEADER
@@ -83,20 +83,21 @@
 import { DateTime } from 'luxon'
 import { trans } from '../common/i18n-renderer'
 import ButtonControl from '../common/vue/form/elements/Button.vue'
+import { defineComponent, PropType } from 'vue'
 
-export default {
+export default defineComponent({
   name: 'CalendarView',
   components: {
     ButtonControl
   },
   props: {
     wordCounts: {
-      type: Object,
-      default: function () { return {} }
+      type: Object as PropType<{ [key: string]: number }>,
+      required: true
     },
     monthlyAverage: {
       type: Number,
-      default: 0
+      required: true
     }
   },
   data: function () {
@@ -114,16 +115,16 @@ export default {
     }
   },
   computed: {
-    year: function () {
+    year: function (): number {
       return this.now.year
     },
-    isCurrentYear: function () {
+    isCurrentYear: function (): boolean {
       return this.now.year === DateTime.local().year
     },
-    calendarLabel: function () {
+    calendarLabel: function (): string {
       return trans('dialog.statistics.tabs.calendar_label')
     },
-    isMinimumYear: function () {
+    isMinimumYear: function (): boolean {
       // Returns true if this.now holds the minimum year for which there is
       // data available
       const years = Object.keys(this.wordCounts).map(k => parseInt(k.substr(0, 4), 10))
@@ -136,7 +137,7 @@ export default {
 
       return this.now.year === min
     },
-    months: function () {
+    months: function (): Array<{ name: string, padding: number, daysInMonth: number }> {
       const ret = []
       const MONTHS = [
         trans('gui.months.jan'),
@@ -165,13 +166,13 @@ export default {
 
       return ret
     },
-    lowMidLegend: function () {
+    lowMidLegend: function (): string {
       return trans('gui.chart.low_mid_legend')
     },
-    highMidLegend: function () {
+    highMidLegend: function (): string {
       return trans('gui.chart.high_mid_legend')
     },
-    highLegend: function () {
+    highLegend: function (): string {
       return trans('gui.chart.high_legend')
     }
   },
@@ -185,16 +186,19 @@ export default {
      *
      * @return  {number}         The percentage from 0 to 1
      */
-    getActivityScore: function (year, month, date) {
-      if (month < 10) {
-        month = `0${month}`
+    getActivityScore: function (year: number, month: number, date: number): number {
+      let parsedMonth = String(month)
+      let parsedDate = String(date)
+
+      if (parsedMonth.length < 2) {
+        parsedMonth = `0${month}`
       }
 
-      if (date < 10) {
-        date = `0${date}`
+      if (parsedDate.length < 2) {
+        parsedDate = `0${date}`
       }
 
-      const wordCount = this.wordCounts[`${year}-${month}-${date}`]
+      const wordCount = this.wordCounts[`${year}-${parsedMonth}-${parsedDate}`]
 
       if (wordCount === undefined || wordCount < this.monthlyAverage / 2) {
         return 0
@@ -206,10 +210,10 @@ export default {
         return 3 // More than twice the monthly average
       }
     },
-    yearMinus: function () {
+    yearMinus: function (): void {
       this.now = this.now.minus({ years: 1 })
     },
-    yearPlus: function () {
+    yearPlus: function (): void {
       // Prevent going into the future
       if (this.now.year === DateTime.local().year) {
         return
@@ -218,7 +222,7 @@ export default {
       this.now = this.now.plus({ years: 1 })
     }
   }
-}
+})
 </script>
 
 <style lang="less">
@@ -263,9 +267,9 @@ body div#calendar-container {
         line-height: 20px;
         font-size: 10px;
 
-        &.low-activity {
-          // Basically no change
-        }
+        // &.low-activity {
+        //   // Basically no change
+        // }
         &.low-mid-activity {
           // Slightly blue-ish
           background-color: @low-mid-bg;